refactor(grpc): error handling
