+++ lib/grpc/GrpcService.ts
@@ -1,18 +1,13 @@
 /* tslint:disable no-floating-promises no-null-keyword */
 import grpc, { status } from 'grpc';
-import Logger from '../Logger';
-import Service from '../service/Service';
-import * as xudrpc from '../proto/xudrpc_pb';
-import { Order, isOwnOrder, OrderPortion, PlaceOrderResult, PlaceOrderEvent, PlaceOrderEventType } from '../orderbook/types';
-import { errorCodes as orderErrorCodes } from '../orderbook/errors';
-import { errorCodes as serviceErrorCodes } from '../service/errors';
-import { errorCodes as p2pErrorCodes } from '../p2p/errors';
-import { errorCodes as swapErrors } from '../swaps/errors';
-import { errorCodes as lndErrorCodes } from '../lndclient/errors';
-import { LndInfo } from '../lndclient/types';
-import { SwapSuccess, SwapFailure } from '../swaps/types';
 import { SwapFailureReason } from '../constants/enums';
-import { TradeInstance, OrderInstance, CurrencyInstance } from '../db/types';
+import { CurrencyInstance, OrderInstance, TradeInstance } from '../db/types';
+import { LndInfo } from '../lndclient/types';
+import { isOwnOrder, Order, OrderPortion, PlaceOrderEvent, PlaceOrderEventType, PlaceOrderResult } from '../orderbook/types';
+import * as xudrpc from '../proto/xudrpc_pb';
+import Service from '../service/Service';
+import { SwapFailure, SwapSuccess } from '../swaps/types';
+import getGrpcError from './getGrpcError';
 
 /**
  * Creates an xudrpc Order message from an [[Order]].
@@ -140,73 +135,7 @@ class GrpcService {
   private streams: Set<grpc.ServerWriteableStream<any>> = new Set<grpc.ServerWriteableStream<any>>();
 
   /** Create an instance of available RPC methods and bind all exposed functions. */
-  constructor(private logger: Logger, private service: Service) {}
-
-  /**
-   * Convert an internal xud error type into a gRPC error.
-   * @param err an error object that should have code and message properties
-   * @return a gRPC error with a gRPC status code
-   */
-  private getGrpcError = (err: any) => {
-    // if we recognize this error, use a proper gRPC ServiceError with a descriptive and appropriate code
-    let code: grpc.status | undefined;
-    switch (err.code) {
-      case serviceErrorCodes.INVALID_ARGUMENT:
-      case p2pErrorCodes.ATTEMPTED_CONNECTION_TO_SELF:
-      case p2pErrorCodes.UNEXPECTED_NODE_PUB_KEY:
-      case orderErrorCodes.MIN_QUANTITY_VIOLATED:
-      case orderErrorCodes.QUANTITY_DOES_NOT_MATCH:
-      case orderErrorCodes.EXCEEDING_LIMIT:
-        code = status.INVALID_ARGUMENT;
-        break;
-      case orderErrorCodes.PAIR_DOES_NOT_EXIST:
-      case p2pErrorCodes.NODE_UNKNOWN:
-      case orderErrorCodes.LOCAL_ID_DOES_NOT_EXIST:
-      case orderErrorCodes.ORDER_NOT_FOUND:
-        code = status.NOT_FOUND;
-        break;
-      case orderErrorCodes.DUPLICATE_ORDER:
-      case p2pErrorCodes.NODE_ALREADY_CONNECTED:
-      case p2pErrorCodes.NODE_ALREADY_BANNED:
-      case p2pErrorCodes.ALREADY_CONNECTING:
-      case orderErrorCodes.CURRENCY_ALREADY_EXISTS:
-      case orderErrorCodes.PAIR_ALREADY_EXISTS:
-        code = status.ALREADY_EXISTS;
-        break;
-      case p2pErrorCodes.NOT_CONNECTED:
-      case p2pErrorCodes.NODE_NOT_BANNED:
-      case p2pErrorCodes.NODE_IS_BANNED:
-      case lndErrorCodes.LND_IS_DISABLED:
-      case orderErrorCodes.CURRENCY_DOES_NOT_EXIST:
-      case orderErrorCodes.CURRENCY_CANNOT_BE_REMOVED:
-      case orderErrorCodes.MARKET_ORDERS_NOT_ALLOWED:
-      case serviceErrorCodes.NOMATCHING_MODE_IS_REQUIRED:
-      case orderErrorCodes.INSUFFICIENT_OUTBOUND_BALANCE:
-      case swapErrors.SWAP_CLIENT_NOT_FOUND:
-        code = status.FAILED_PRECONDITION;
-        break;
-      case lndErrorCodes.LND_IS_UNAVAILABLE:
-      case p2pErrorCodes.COULD_NOT_CONNECT:
-        code = status.UNAVAILABLE;
-        break;
-      case p2pErrorCodes.POOL_CLOSED:
-        code = status.ABORTED;
-        break;
-      case p2pErrorCodes.RESPONSE_TIMEOUT:
-        code = status.DEADLINE_EXCEEDED;
-        break;
-    }
-
-    // return a grpc error with the code if we've assigned one, otherwise pass along the caught error as UNKNOWN
-    const grpcError: grpc.ServiceError = {
-      code: code || status.UNKNOWN,
-      message: err.message,
-      name: err.name,
-    };
-    this.logger.error(grpcError);
-
-    return grpcError;
-  }
+  constructor(private service: Service) {}
 
   /** Closes and removes all active streaming calls. */
   public closeStreams = () => {
@@ -234,7 +163,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -248,7 +177,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -262,7 +191,7 @@ class GrpcService {
       response.setQuantityOnHold(quantityOnHold);
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -282,7 +211,7 @@ class GrpcService {
       });
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -295,7 +224,7 @@ class GrpcService {
       const response = new xudrpc.OpenChannelResponse();
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -309,7 +238,7 @@ class GrpcService {
       const response = new xudrpc.ConnectResponse();
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -322,7 +251,7 @@ class GrpcService {
       const response = new xudrpc.BanResponse();
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -335,7 +264,7 @@ class GrpcService {
       const response = new xudrpc.UnbanResponse();
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -377,10 +306,9 @@ class GrpcService {
           name: SwapFailureReason[err],
           message: SwapFailureReason[err],
         };
-        this.logger.error(grpcError);
         callback(grpcError, null);
       } else {
-        callback(this.getGrpcError(err), null);
+        callback(getGrpcError(err), null);
       }
     }
   }
@@ -454,7 +382,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -471,7 +399,7 @@ class GrpcService {
       response.setReputationscore(reputationScore);
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -500,7 +428,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -523,7 +451,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -538,7 +466,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -565,7 +493,7 @@ class GrpcService {
       response.setTradesList(tradesList);
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -597,7 +525,7 @@ class GrpcService {
       response.setPeersList(peers);
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -612,7 +540,7 @@ class GrpcService {
 
       call.end();
     } catch (err) {
-      call.emit('error', this.getGrpcError(err));
+      call.emit('error', getGrpcError(err));
     }
   }
 
@@ -624,7 +552,7 @@ class GrpcService {
       const result = await this.service.placeOrder(call.request.toObject());
       callback(null, createPlaceOrderResponse(result));
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -638,7 +566,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -652,7 +580,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -668,7 +596,7 @@ class GrpcService {
 
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
 
@@ -678,7 +606,7 @@ class GrpcService {
       const response = new xudrpc.ShutdownResponse();
       callback(null, response);
     } catch (err) {
-      callback(this.getGrpcError(err), null);
+      callback(getGrpcError(err), null);
     }
   }
